backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
redis==5.0.1

# Data & Analysis
pandas==2.2.0
numpy==1.26.3
yfinance==0.2.35
ta-lib==0.4.28; platform_system != "Windows"

# Utils
pydantic==2.5.3
python-multipart==0.0.6
